[bots][infra] Renaming `generate` to `snapshot`
This PR renames `bot.py generate` to `bots.py snapshot`. This is being
done to avoid confusion down the line when `bots.py gen` is introduced.

This change also splits `snapshot` out of `bots.py`, the same way
`lookup` has been placed out of it.

Bug: https://github.com/brave/brave-browser/issues/47912
